WebDECIMER/auto.requirements.txt
absl-py==0.11.0
appnope==0.1.0
asgiref==3.3.1
astor==0.8.1
astunparse==1.6.3
backcall==0.2.0
cached-property==1.5.2
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
Django==3.1.3
django-cors-headers==3.5.0
django-filter==2.4.0
django-storages==1.10.1
djangorestframework==3.12.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
idna==2.10
imageio==2.9.0
imantics==0.1.12
importlib-metadata==2.0.0
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
joblib==0.17.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lxml==4.6.1
Markdown==3.3.3
matplotlib==3.3.3
networkx==2.5
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
pandas==1.1.4
parso==0.7.1
pdf2image==1.14.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pip-chill==1.0.0
prompt-toolkit==3.0.8
protobuf==3.14.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.7.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sqlparse==0.4.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.10.1
traitlets==5.0.5
urllib3==1.26.2
uWSGI==2.0.19.1
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
xmljson==0.2.1
zipp==3.4.0
